# Increase orderbook mid price cache duration to 60 seconds