Increases the amount of time we store orderbook mid price data from 30 to 60 seconds. This is to remedy spikes in candle data that we continue to see with 30 seconds of data. 

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **Bug Fixes**
	- Adjusted the timing in the `getMedianPrice` test case from 30 seconds to 60 seconds to ensure accurate median price verification.

- **Chores**
	- Updated the time window for processing market prices from 30 seconds to 60 seconds in the relevant Lua script, including necessary comments.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->