@adamfraser 2024-12-10
@mergify backport release/indexer/v8.x