@adamfraser 2024-12-10
@mergify backport release/indexer/v7.x